requirements.txt
pyarrow
tqdm
skimpy
astronomer-cosmos[dbt-bigquery]
soda-core-bigquery
apache-airflow-providers-google
google-cloud-bigquery-storage
protobuf
pylance